Specify --as-needed while generating shared library
The use of the --as-needed flag allows the linker to avoid linking extra
libraries in a binary. This not only improves startup times (as the
loader does not have to load all the libraries for every step) but might
avoid the full initialization.

More importantly, the use of --as-needed avoids adding dependencies to a
binary that are prerequisites of one of its direct or indirect
dependencies. This is important because when a library changes SONAME
after an ABI change, all the binaries directly linking to it have to be
rebuilt. By linking only the libraries that are actually needed, the
breakage due to an ABI change is reduced.